requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
langgraph-cli[inmem]
pytest-playwright